+++ lib/generators/react_email_rails/templates/email/component.tsx
@@ -1,4 +1,4 @@
-import { Body, Container, Heading, Html, Text } from "@react-email/components"
+import { Body, Container, Heading, Html, Text } from "react-email"
 
 export default function <%= @component_name %>() {
   return (